release: support tms cloud. (#48)

+++ src/main/java/ru/testit/client/model/UpdateWorkItemApiModel.java
@@ -98,7 +98,7 @@ public class UpdateWorkItemApiModel {
 
   public static final String JSON_PROPERTY_DURATION = "duration";
   @jakarta.annotation.Nonnull
-  private Integer duration;
+  private Long duration;
 
   public static final String JSON_PROPERTY_ATTRIBUTES = "attributes";
   @jakarta.annotation.Nonnull
@@ -334,7 +334,7 @@ public class UpdateWorkItemApiModel {
   }
 
 
-  public UpdateWorkItemApiModel duration(@jakarta.annotation.Nonnull Integer duration) {
+  public UpdateWorkItemApiModel duration(@jakarta.annotation.Nonnull Long duration) {
     this.duration = duration;
     return this;
   }
@@ -349,14 +349,14 @@ public class UpdateWorkItemApiModel {
   @JsonProperty(JSON_PROPERTY_DURATION)
   @JsonInclude(value = JsonInclude.Include.ALWAYS)
 
-  public Integer getDuration() {
+  public Long getDuration() {
     return duration;
   }
 
 
   @JsonProperty(JSON_PROPERTY_DURATION)
   @JsonInclude(value = JsonInclude.Include.ALWAYS)
-  public void setDuration(@jakarta.annotation.Nonnull Integer duration) {
+  public void setDuration(@jakarta.annotation.Nonnull Long duration) {
     this.duration = duration;
   }
